@@ -2,6 +2,7 @@
   "name": "my-app",
   "version": "0.1.0",
   "private": true,
+  "type": "module",
   "dependencies": {
     "airtable": "^0.8.1",
     "axios": "^0.21.1",
@@ -11,26 +12,26 @@
     "gridlex": "^2.7.1",
     "http-parser-js": "^0.5.3",
     "node-cron": "^3.0.0",
-    "node-fetch": "^3.0.0-beta.9",
-    "node-sass": "^4.14.1",
+    "node-fetch": "^2.6.6",
+    "node-sass": "^8.0",
     "onkyo.js": "^0.8.0",
-    "react": "^17.0.2",
+    "react": "^18.0.0",
     "react-best-tabs": "^1.1.2",
-    "react-dom": "^17.0.2",
-    "react-native": "^0.64.2",
+    "react-dom": "^18.0.0",
+    "react-native": "^0.71",
     "react-router-dom": "^4.3.1",
     "react-scripts": "^4.0.3",
     "react-toggle": "^4.1.1",
     "reactstrap": "8.9",
-    "socket.io": "^2.3.0",
-    "socket.io-client": "^2.2.0",
+    "socket.io": "^4.6.0",
+    "socket.io-client": "^4.6.0",
     "styled-components": "^5.1.0",
     "xmlhttprequest": "^1.8.0"
   },
   "scripts": {
-    "start": "react-scripts start",
+    "start": "react-scripts --openssl-legacy-provider start",
     "server-build": "rsync -r ./src/services/* ./build/services && rsync -r ./src/environment.js ./build",
-    "build": "react-scripts build && npm run server-build",
+    "build": "react-scripts --openssl-legacy-provider build && npm run server-build",
     "socket-serve--dev": "node src/services/socket.server.js &",
     "dev": "npm install && npm start",
     "test": "npm run socket-serve--dev && npm start"
